linux: Assign default info for generic kernel
This clears the way for the custom-kernel-info to reside in a
conf/ directory, which can be defined by whoever is setting up
OE and Bitbake.

Therefore, instead of setting the custom kernel information
and placing it inside the layer (Git controlled), that file
can now reside under the build conf/, along with the
lkft-kernel.conf and other local configurations.

Nothing wrong happens if the file doesn't exist because:
1) it's included, not required, and
2) defaults are assigned in the kernel recipe.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>