classes/kernel: Consider {Modules,vmlinux}.symvers change from next-20200528
As of next-20200528, specifically because of f04dad545fa6
("modpost: generate vmlinux.symvers and reuse it for the
second modpost"), the Modules.symvers file is not necessarily
there when we want to copy it in the kernel bbclass.

With this change here, we check if it exists before copying it
and also consider the new vmlinux.symvers file.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>

(Cherry-picked from commit 8b1c93fa10b4299b50a9ee827dbc1a8a841e9817)