linux: Fix missing cflags for host compilation
While building the scripts/ directory of the kernel, a problem
occurs when openssl/bio.h cannot be found:
|   HOSTCC  scripts/extract-cert
| /poky/build/tmp/work-shared/machine/kernel-source/scripts/extract-cert.c:21:10: fatal error: openssl/bio.h: No such file or directory
|  #include <openssl/bio.h>
|           ^~~~~~~~~~~~~~~
| compilation terminated.
| make[2]: *** [scripts/Makefile.host:103: scripts/extract-cert] Error 1
| make[2]: *** Waiting for unfinished jobs....
| make[1]: *** [/poky/build/tmp/work-shared/machine/kernel-source/Makefile:1133: scripts] Error 2
| make: *** [/poky/build/tmp/work-shared/machine/kernel-source/Makefile:180: sub-make] Error 2
| WARNING: exit code 1 from a shell command.

That header file, however, is correctly placed in the sysroot
include directory. This was probably not a problem before
since we always build where the libssl-dev package is
already installed, but that's the host's version of the
library, not the one we just built.

By exporting the variable we make sure that the kernel's
Makefile is able to get that definition, so that the
right include is set.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>